refactor: extrac authentication to seperate module
